swe_agent_system/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "swe-agent-system"
version = "0.1.0"
description = "Enterprise-Grade Multi-Agent Software Engineering AI System"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SWE Agent Team"}
]
keywords = ["ai", "agents", "software-engineering", "automation", "qiskit"]

dependencies = [
    "google-genai>=1.0.0",
    "PyGithub>=2.1.0",
    "pyyaml>=6.0",
    "structlog>=23.2.0",
    "pydantic>=2.5.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
swe-agent = "swe_agent_system.cli:main"

[tool.setuptools.packages.find]
where = [".."]
include = ["swe_agent_system*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
